+++ python/ray/data/datasource/image_datasource.py
@@ -43,6 +43,8 @@ class ImageDatasource(FileBasedDatasource):
 
     _WRITE_FILE_PER_ROW = True
     _FILE_EXTENSION = ["png", "jpg", "jpeg", "tif", "tiff", "bmp", "gif"]
+    # Use 8 threads per task to read image files.
+    _NUM_THREADS_PER_TASK = 8
 
     def create_reader(
         self,
